requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitarray==2.6.0
boto3==1.24.36
botocore==1.27.36
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cookiecutter==2.1.1
cryptography==37.0.4
cytoolz==0.12.0
Deprecated==1.2.13
Django==4.0.3
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-csp==3.7
django-encrypted-model-fields==0.6.5
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-softdelete==0.9.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-jwt==1.19.2
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Faker==13.15.1
frozenlist==1.3.1
gcloud==0.18.3
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
googlemaps==4.6.0
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
hexbytes==0.3.0
httplib2==0.20.4
idna==3.3
imagekitio==2.2.8
importlib-metadata==4.12.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
jsonschema==4.14.0
jwcrypto==1.3.1
jws==0.1.3
langdetect==1.0.9
lru-dict==1.1.8
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.2
mysql-connector-python==8.0.29
mysqlclient==2.1.1
netaddr==0.8.0
numpy==1.23.3
oauth2client==4.1.3
openai==0.23.1
openpyxl==3.0.10
pandas==1.5.0
pandas-stubs==1.5.0.221003
parsimonious==0.8.1
poyo==0.5.0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyfcm==1.5.4
PyJWT==2.2.0
pyngrok==5.1.0
pyparsing==3.0.9
Pyrebase4==4.5.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jwt==3.3.3
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rlp==2.0.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==4.1.0
text-unidecode==1.3
toolz==0.12.0
tqdm==4.64.1
types-pytz==2022.4.0.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.10
varint==1.0.2
web3==5.30.0
websockets==9.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
